chore: refactor my collection screen logic
