+++ src/app/Scenes/MyCollection/MyCollectionArtworks.tsx
@@ -1,13 +1,9 @@
-import { ActionType, AddCollectedArtwork, ContextModule, OwnerType } from "@artsy/cohesion"
-import { Spacer, LockIcon, Flex, useSpace, Text, Button } from "@artsy/palette-mobile"
+import { Button, Flex, Spacer, Text } from "@artsy/palette-mobile"
 import { MyCollection_me$data } from "__generated__/MyCollection_me.graphql"
 import { ArtworksFiltersStore } from "app/Components/ArtworkFilter/ArtworkFilterStore"
 import { FilteredArtworkGridZeroState } from "app/Components/ArtworkGrids/FilteredArtworkGridZeroState"
 import { InfiniteScrollMyCollectionArtworksGridContainer } from "app/Components/ArtworkGrids/InfiniteScrollArtworksGrid"
-import { ZeroState } from "app/Components/States/ZeroState"
-import { Tab } from "app/Scenes/MyProfile/MyProfileHeaderMyCollectionAndSavedWorks"
 import { GlobalStore } from "app/store/GlobalStore"
-import { navigate, popToRoot } from "app/system/navigation/navigate"
 import { cleanLocalImages } from "app/utils/LocalImageStore"
 import { extractNodes } from "app/utils/extractNodes"
 import { useScreenDimensions } from "app/utils/hooks"
@@ -16,15 +12,13 @@ import { refreshMyCollection } from "app/utils/refreshHelpers"
 import { useEffect, useState } from "react"
 import {
   Alert,
-  Image,
   InteractionManager,
   LayoutAnimation,
   NativeScrollEvent,
   NativeSyntheticEvent,
 } from "react-native"
 
-import { graphql, RelayPaginationProp } from "react-relay"
-import { useTracking } from "react-tracking"
+import { RelayPaginationProp, graphql } from "react-relay"
 import { MyCollectionArtworkList } from "./Components/MyCollectionArtworkList"
 import { MyCollectionSearchBar } from "./Components/MyCollectionSearchBar"
 import { MyCollectionArtworkEdge } from "./MyCollection"
@@ -69,10 +63,6 @@ export const MyCollectionArtworks: React.FC<MyCollectionArtworksProps> = ({
   useEffect(() => {
     cleanLocalImages()
   }, [])
-
-  if (artworks.length === 0) {
-    return <MyCollectionZeroState />
-  }
 
   // Make Search Bar visible when user scrolls to top
   const handleScroll = ({ nativeEvent }: NativeSyntheticEvent<NativeScrollEvent>) => {
@@ -197,56 +187,6 @@ export const MyCollectionArtworks: React.FC<MyCollectionArtworksProps> = ({
   )
 }
 
-const MyCollectionZeroState: React.FC = () => {
-  const { trackEvent } = useTracking()
-  const space = useSpace()
-
-  const image = require("images/my-collection-empty-state.webp")
-
-  return (
-    <ZeroState
-      bigTitle="Know Your Collection Better"
-      subtitle="Manage your collection online and get free market insights."
-      image={
-        <Image
-          source={image}
-          resizeMode="contain"
-          style={{
-            alignSelf: "center",
-            marginVertical: space(2),
-          }}
-        />
-      }
-      callToAction={
-        <>
-          <Button
-            testID="add-artwork-button-zero-state"
-            onPress={() => {
-              trackEvent(tracks.addCollectedArtwork())
-              navigate("my-collection/artworks/new", {
-                passProps: {
-                  mode: "add",
-                  source: Tab.collection,
-                  onSuccess: popToRoot,
-                },
-              })
-            }}
-            block
-          >
-            Upload Artwork
-          </Button>
-          <Flex flexDirection="row" justifyContent="center" alignItems="center" py={1}>
-            <LockIcon fill="black60" />
-            <Text color="black60" pl={0.5} variant="xs">
-              My Collection is not shared with sellers.
-            </Text>
-          </Flex>
-        </>
-      }
-    />
-  )
-}
-
 /**
  * * IMPORTANT *
  *
@@ -286,12 +226,3 @@ export const MyCollectionFilterPropsFragment = graphql`
     }
   }
 `
-
-const tracks = {
-  addCollectedArtwork: (): AddCollectedArtwork => ({
-    action: ActionType.addCollectedArtwork,
-    context_module: ContextModule.myCollectionHome,
-    context_owner_type: OwnerType.myCollection,
-    platform: "mobile",
-  }),
-}
